% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mtt_calc.R
\name{mtt_calc}
\alias{mtt_calc}
\title{Calculate normalized ODs without fitting}
\usage{
mtt_calc(
  data,
  signal = "nm562",
  background = "nm660",
  dose = "dose",
  out = "div",
  .by = NULL
)
}
\arguments{
\item{data}{A \code{data.frame} of absorbances, containing at least the \code{signal},
\code{background}, and \code{dose} columns (see below)}

\item{signal}{The name of the column that contains the signal absorbances}

\item{background}{The name of the column that contains the background
absorbances. If NULL, will just use signal.}

\item{dose}{The name of the column containing the concentrations of the drug.
Will use the lowest concentration as the baseline.}

\item{out}{The name of the output column}

\item{.by}{Character vector of columns to summarize by. Other columns will be
dropped}
}
\description{
It can be useful to calculate normalized ODs without fitting, particularly if
you would like to use \code{geom_mtt}, or need to manipulate your data downstream.
}
