% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mtt_fit.R
\name{mtt_fit}
\alias{mtt_fit}
\alias{mtt_fit.default}
\alias{mtt_fit.gp}
\alias{mtt_fit.data.frame}
\alias{mtt_fit.spectramax}
\title{Fit an curve through MTT data}
\usage{
mtt_fit(x, ...)

\method{mtt_fit}{default}(x, ...)

\method{mtt_fit}{gp}(x, ...)

\method{mtt_fit}{data.frame}(x, ...)

\method{mtt_fit}{spectramax}(x, conditions, ...)
}
\arguments{
\item{x}{An object of class \code{gplate::gp}, \code{data.frame}, or \code{mop::spectramax}.
See details.}

\item{...}{Unused}

\item{conditions}{A named list of numerics of length 4. Contains doses per
quadrant. Names are the drug. If quadrant should be omitted, supply NA.
Quadrants go from left to right, top to bottom. Identical names are
allowed.}
}
\description{
Fit an curve through MTT data
}
\details{
If a \code{data.frame} or \code{gplate::gp} is supplied, it should have
columns \code{condition}, \code{dose}, \code{nm562} and \code{nm660}
}
