% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mtt_plot.R
\name{mtt_plot}
\alias{mtt_plot}
\title{Plot MTT fits}
\usage{
mtt_plot(fits, ic_pct = NULL)
}
\arguments{
\item{ic_pct}{Numeric. If NULL, no IC will be plotted. Otherwise, will plot
\code{ic_pct}\% IC the \% IC supplied.}

\item{mtt}{a list of fits, usually the output from \code{mtt_fit}}
}
\value{
a \code{ggplot}
}
\description{
Plot MTT fits
}
