% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mtt_tidy.R
\name{mtt_tidy}
\alias{mtt_tidy}
\title{Tidy mtt data}
\usage{
mtt_tidy(x, ...)
}
\arguments{
\item{x}{Object to be tidied}

\item{...}{Additional arguments passed to appropriate methods}

\item{conditions}{A named list of length 4, where the name is the name of the
condition, and the values are the doses. If a quadrant of the plate was not
used, use NA.}
}
\description{
A convenience function for tidying mtt data. For data that do not follow the
standard format of 'one condition per quadrant', consider tidying the data
using \code{gplate}. See the "Using ezmtt" vignette for more information.
}
