% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mtt_plot.R
\name{mtt_plot}
\alias{mtt_plot}
\title{Plot MTT results}
\usage{
mtt_plot(fits, plot_ics = FALSE)
}
\arguments{
\item{plot_ics}{logical. Should the calculated inhibitor concentrations be
plotted?}

\item{mtt}{a \code{data.frame} output from  \code{mtt_calc()}}
}
\value{
a \code{ggplot}
}
\description{
Plot MTT results
}
